cs 6241 project - update system of ODEs to 3D
Reduced dimension of state space from 4D to 3D by considering the R-type mosquito population as a whole instead of separate male/female subentities. Three ODEs considered are now R, Fw, and Mw. This will make it more straightforward for me to write a solver for the Hamilton Jacobi Bellman equation that we can use to compare our results to.