zika - update cpp hjb solver
Updated the boundary conditions. Playing around with initial conditions, and there are definitely spots in which the controls are nonzero!

I haven't thought of the best way to visualize the controls yet. The jupyter plotting notebook I included now doesn't have any commentary yet, and the "coloring point in 3D space according to the optimal control value" is NOT the way to go here (as shown!) since it's far too dense to see anything valuable. I'll be working on this over the weekend more. The first step for me will likely be to write a "trajectory tracer" which would plot the optimal trajectory for one set of initial conditions. This would be good for visualization, but unfortunately it doesn't really give us a global / overall sense of what's going on with the controls.